ARROW-6754: [C++] Merge allocator.h into stl.h
